pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pelix"]

[project]
name = "iPOPO"
version = "3.1.0"
authors = [
  { name = "Thomas Calmant", email = "thomas.calmant+github@gmail.com" },
]
requires-python = ">=3.10"
description = "A service-oriented component model framework"
readme = "README.rst"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]
license = "Apache-2.0"
license-files = ["LICENSE"]
dependencies = ["jsonrpclib-pelix>=0.4.3"]

[project.urls]
Homepage = "https://ipopo.readthedocs.io/"
Documentation = "https://ipopo.readthedocs.io/"
Issues = "http://github.com/tcalmant/ipopo/issues"
Source = "http://github.com/tcalmant/ipopo/"

[project.optional-dependencies]
etcd2 = ["python-etcd==0.4.5", "osgiservicebridge>=1.5.8"]
MQTT = ["paho-mqtt>=2.1"]
Redis = ["redis>=2.10"]
RSA = ["osgiservicebridge>=1.5.8", "grpcio>=1.71.0", "grpcio-tools>=1.71.0"]
XMPP = ["slixmpp==1.10"]
zeroconf = ["zeroconf==0.19"]
ZooKeeper = ["kazoo==2.8.0"]
yaml = ["pyyaml>=6.0"]

[tool.black]
line-length = 110
